// backend/src/entities/enums.ts
export enum UserRole {
  USER = 'user',
  ADMIN = 'admin',
}

export enum TransactionType {
  INCOME = 'income',
  EXPENSE = 'expense',
  TRANSFER = 'transfer',
}

export enum AccountType {
  CASH = 'cash',
  BANK = 'bank',
  CARD = 'card',
  SAVINGS = 'savings',
  CREDIT = 'credit',
}

export enum Currency {
  AMD = 'AMD',
  USD = 'USD',
  EUR = 'EUR',
}

export enum BudgetPeriod {
  WEEKLY = 'weekly',
  MONTHLY = 'monthly',
  YEARLY = 'yearly',
}